Clean up StorybookLoader.
