+++ packages/core/config/storybook/preview.js
@@ -1,11 +1,13 @@
 const React = require('react')
 const { addDecorator } = require('@storybook/react')
 
+// The StorybookLoader is responsible for importing all the mock files and booting
+// up the mock service workers.
+const { StorybookLoader } = require('@redwoodjs/core/dist/storybook')
+
 // Import the user's default CSS file
 require('~__REDWOOD__USER_WEB_DEFAULT_CSS')
 
-// The StorybookLoader is responsible for importing the mock files and booting
-// up the mock service workers.
-const { default: StorybookLoader } = require('@redwoodjs/core/dist/StorybookLoader')
-
-addDecorator((storyFn) => React.createElement(StorybookLoader, { storyFn }))
+addDecorator(
+  (storyFn) => React.createElement(StorybookLoader, { storyFn })
+)
